+++ GameEngine/src/renderer/Camera.cpp
@@ -23,7 +23,7 @@ void Camera::CalcViewProjectionMatrix()
 	//	* glm::lookAt(glm::vec3(m_GlobalTransform[3]), glm::vec3(m_GlobalTransform[3]) + front, up);
 
 	m_ViewMatrix = glm::lookAt(glm::vec3(m_GlobalTransform[3]), glm::vec3(m_GlobalTransform[3]) + front, up);
-	m_ProjectionMatrix = glm::perspective<float>(glm::radians(m_FOV), Window::g_WindowWidth / Window::g_WindowHeight, m_NearPlane, m_FarPlane);	// TODO: make callback instead of calc every frame
+	m_ProjectionMatrix = glm::perspective<float>(glm::radians(m_FOV), (float)Window::g_WindowWidth / Window::g_WindowHeight, m_NearPlane, m_FarPlane);	// TODO: make callback instead of calc every frame
 }
 
 // PRIVATE
